move: util 패키지 하위에 api 패키지 생성
list
- ApiResponse
- ResponseMessage

추가로 메세지 끝의 온점을 제거했습니다.
