+++ src/main/java/site/onandoff/auth/presentation/AuthController.java
@@ -1,6 +1,6 @@
 package site.onandoff.auth.presentation;
 
-import static site.onandoff.util.ResponseMessage.*;
+import static site.onandoff.util.api.ResponseMessage.*;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,8 +13,8 @@ import site.onandoff.auth.application.AuthService;
 import site.onandoff.auth.dto.AuthenticationTokenPair;
 import site.onandoff.auth.dto.LoginData;
 import site.onandoff.auth.dto.ReissuedAccessToken;
-import site.onandoff.util.ApiResponse;
 import site.onandoff.util.Login;
+import site.onandoff.util.api.ApiResponse;
 
 @RestController
 @RequiredArgsConstructor
